+++ src/lib/components/modes/textmode/StatusBar.scss
