+++ src/lib/components/modes/textmode/StatusBar.svelte
