+++ src/lib/components/modes/textmode/TextMode.scss
